Refactor UI a bit for better organization
